get temporary extraction directory from env
use '$XDG_CACHE_HOME/mcomix' as the extraction dir with fallback
set to '/tmp/mcomix'

removed ability to set tmp dir

the smart thing to do is having $XDG_CACHE_HOME be set on a tmpfs since
everything writen there is junk anyway. not sure about the fallback dir
location but realistically it does not even matter since it should never
be used.