change default number of pages to cache
Using a page cache of 3 forward and 2 backwards, this allows for the
next page change to have both pages preloaded. A higher number has
little benefit. PAGE_CACHE_FORWARD has a minumum cache of one because
when showing two pages you need page+1 to be loaded.

From a memory usage perspective a lower cache amount is better since
a GtkPixbuf will use a lot of memory relative to the image size.