pybind11 Box() / Layout() / Zoom() impl
This imports cpp image layout code.

This code is from the cpp rewrite I am working on. These classes are
independent from the GTK code and responsible for image sizing, layout,
and posistioning. Since one of the current issues with the rewrite
involves the layout and sizing of the Gtk::Image's in a Gtk::Fixed in a
Gtk::ScrolledWindow, I wanted reassurances that it was a problem with
the GTK code and not the layout code.

These new versions are basiclly identical to their python version, with
tests that run and pass on both implementations. Their output should be
identical.

TODO
ZoomModel::_calc_limits() requires access to the config if
the arg fitmode is ZoomModes.SIZE.