Port from python and GTK3 to C++ and GTK4.
This implements all of the important features.

Image loading is still done with the deprecated Gtk::Pixbuf class
since there is not an equivalent replacement. Tried using a
Gdk::Texture in a Gtk::Snapshot but that just looked bad.

Missing features to be added later
- page selector
- bookmarks,
- preferences dialog
- properties dialog
- file open dialog
- zoom levels

This has been a project I have been working on and off again, mostly
off, for a while.

The main reason this took so long for me to finish was getting the
image layout to work. In the python version the images are arranged
in a Gtk::Layout, that was removed in GTK4 and Gtk::Fixed was
recommended as the replacement. I could never get Gtk::Fixed to
properly scale and place the images or deal with window resize events
correctly which caused this project to essentially get abandoned.

KNOWN BUGS:
- rotation 90 and 270 shows images side by side not top to bottom,
not really a bug just the result of the new layout handling
- thumbbar a page change selects an item in the selection model but
it does not visibly select the item
- thumbbar the idle handler does not get cleared when changing
archives before all thumbnails have loaded resulting in extra
thumbnails from the old archive being in the sidebar for the archive
being opened.
-thumbbar if there is a wide thumbnail in the list normal page width
thumbnails do not get aligned in the center