remove settings.open_first_page
open_first_page was used in the python version to start at either the
first or last page when opening a previous archive. this code never
worked here since it was trying to use python style list indexing.

If you want the last page when opening a previous archive just press
the End key after it opens.