simple crash recovery
this only exists because my window manager crashed once 2 years ago
and it annoyed me enough that I put this on the TODO list for the
python version but never got around to implementing it.

The list/recovery commands should probably get moved into a seperate
binary in tools/

this also adds the original low tech recovery script.